services/mcp-server/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
pydantic-settings==2.2.1
asyncpg==0.29.0
qdrant-client==1.9.0
redis==5.0.3
aiohttp==3.9.5
python-multipart==0.0.9
prometheus-fastapi-instrumentator==6.1.0
httpx==0.27.0
pandas==2.2.0
sentence-transformers==2.6.1
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
duckduckgo-search==5.3.0
google-api-python-client==2.117.0
tavily-python==0.3.3
serpapi==0.1.5
pdfplumber==0.10.4
